refactor(tof): fold flexible rebinning into rebin_tof, retire flexible_rebinner module (#192)
Honors the Q5 "extend existing rebinning" agreement: #192 had shipped a separate
flexible_rebinner module with a parallel public API (apply_tof_rebin, rebin_tof_by_list)
that merely wrapped rebin_tof. This consolidates everything into histogram_rebinner.rebin_tof
as the single entry point.

- rebin_tof gains reduction=None|"sum"|"mean"|"median" and accepts an explicit
  [[start, stop], ...] frame-index bin list as its width. Sum (None/"sum" + scalar width)
  keeps the existing scipp.rebin adjacent-bin behavior verbatim (backward compatible);
  "mean"/"median", or a bin list, take the frame-index reduce path (gaps -> dropped_frames
  mask + NaN + spectra_tof). mean/median reject non-bins unit + logarithmic (scipp.rebin is
  sum-only) with a clear error directing to log_bin_list.
- reduce_tof_bins, linear_bin_list, log_bin_list + the SPECTRA_TOF_COORD / DROPPED_FRAMES_MASK
  constants move into histogram_rebinner as public helpers. apply_tof_rebin and rebin_tof_by_list
  are retired; flexible_rebinner.py is deleted.
- The 3 VENUS TOF pipelines call rebin_tof directly, with an inlined spec-type guard preserving
  the "bool, an int factor, or a list" error for invalid rebin_by_tof.
- Docs (api.rst autodoc, CHANGELOG, workflow guides) point at rebin_tof; test_flexible_rebinner
  renamed to test_rebin_tof_flexible.

No behavior change to the shipped feature. pixi run test -> 596 passed; pre-commit clean;
build-docs succeeded.

Assisted-With: claude-opus-4-8
