+++ docs/api.rst
@@ -75,9 +75,6 @@ Time-of-flight
 .. automodule:: neunorm.tof.histogram_rebinner
    :members:
 
-.. automodule:: neunorm.tof.flexible_rebinner
-   :members:
-
 .. automodule:: neunorm.tof.pixel_detector
    :members:
